refactor: move dataview lookup helpers to cli/commands/discovery.py
